Update lensing.py
Changed it back to using the second to last position & velocity. Shouldn't make much of a difference for what we've done so far, but as you get closer to the black hole, using the final velocity from the trajectory will be more significantly different from using the second to final velocity.